+++ driver-core/CHANGELOG.rst
@@ -12,6 +12,7 @@ CHANGELOG
 - [bug] Fix potential NPE in HostConnectionPool (JAVA-231)
 - [bug] Make sure connections are returned to the right pools (JAVA-255)
 - [bug] Use date string in query build to work-around CASSANDRA-6718 (JAVA-264)
+- [bug] Always return the PreparedStatement object that is cache internally (JAVA-263)
 
 
 1.0.5: